finish xml table
