+++ magda-web-client/src/actions/previewDataActions.js
@@ -3,8 +3,7 @@ import fetch from 'isomorphic-fetch'
 import {actionTypes} from '../constants/ActionTypes';
 import parser from 'rss-parser'
 import papa from 'papaparse';
-import  fastXmlParser from 'fast-xml-parser';
-import traverse from 'traverse';
+import xmlToTabular from '../helpers/xmlToTabular';
 
 
 export function requestPreviewData(fileName){
@@ -74,40 +73,16 @@ export function fetchPreviewData(distributions){
       }else if(format === 'xml'){
         fetch(url)
         .then(response=>
-          response.text()
+          {
+            if (response.status !== 200) {return dispatch(requestPreviewDataError(response.status))}
+            return response.text();
+          }
         ).then(xmlData=>{
-          // when a tag has attributes
-            var options = {
-                attrPrefix : "@_",
-                textNodeName : "#text",
-                ignoreNonTextNodeAttr : true,
-                ignoreTextNodeAttr : true,
-                ignoreNameSpace : true,
-                ignoreRootElement : false,
-                textNodeConversion : true,
-                textAttrConversion : false
-            };
-            if(fastXmlParser.validate(xmlData)=== true){//optional
-            	const jsonObj = fastXmlParser.parse(xmlData,options);
-              var array = traverse(jsonObj).reduce(function (acc) {
-                  if (this.notRoot && this.isLeaf) {
-                    acc.push(
-                      {
-                        name:this.parent.key,
-                        value:this.node
-                      }
-                    );
-                  }
-                  return acc;
-              }, []);
-                const data = {
-                  data: array,
-                  meta: {
-                    fields: ['name', 'value']
-                  }
-                }
-                return dispatch(receivePreviewData(data))
-            }
+          if(xmlToTabular(xmlData)){
+            dispatch(receivePreviewData(xmlToTabular(xmlData)));
+          } else{
+            dispatch(requestPreviewDataError('failed to parse xml'))
+          }
         })
       }
   }
